updated main.js
updated the selectors in the js to fix the document

i was so close. i was so unfathomably close that it hurts. i spent so long trying to find some archaic solution on the darkest recesses of the web and it turned out i was just using the wrong selectors are you joking aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa